Create engine_scripting_example.scene
Example scene file for Sprint 2 compatible code.
It should be noted that this code DOES NOT provide a complete scene description. It simply illustrates the way nodes and script are implemented in scene files. To make a complete scene file with the above example, simply copy and paste this code into your scene file or write the global scene data around this code.
Also, I wrote this without a reference, so you may have to change some of the variable names, values, etc.